+++ changelog.d/20230511_105948_sirosen_allow_flows_list_multi_orderby.rst
@@ -1,0 +1,2 @@
+* ``FlowsClient.list_flows`` now supports passing a non-string iterable of
+  strings to ``orderby`` in order to indicate multiple orderings (:pr:`NUMBER`)
